Moved tests from subfolders into main folders.
